moved code to stroke module
