+++ stroke/__init__.py
