+++ stroke/bids_loader.py
