+++ stroke/indi_reformat.py
