+++ stroke/nii_slice.py
