+++ stroke/prediction.py
