+++ stroke/scoring.py
