+++ stroke/setup.py
