+++ stroke/stroke_config.py
